secanomalyd: adds user namespace check for unpriv
Adds a check for the user namespace when evaluating the percentage of
unprivileged processes. Updates the condition such that if the process
is not root AND (it has no CAP_SYS_ADMIN OR it is in a different user
namespace than the init process), it is considered unprivileged.
Converts all instances of const ino_t& to ino_t as this is OK to pass by
value.

BUG=b:293617627
TEST=All unit tests pass. Percentage of unprivileged processes drops
slightly.

Change-Id: I63fa8628d6a24c168bda6e22506421a9452be605
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4729358
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>