cryptohome: Another approach to typestates in AuthSession
After abandoning the first expriment, this change takes a different
approach, replacing the list of authorized intents with a set of
subobjets, one per intent. So the idea is that operations that require a
specific intent (generally, "decrypt") can only be accessed by querying
for and checking for that specific intent object.

This change goes a bit further than the other prototype in that it
doesn't just use these subobjects internally but actually makes them
visible in the public API.

BUG=b:274971903
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Id28a32a46ac5e537757dd623bc01f6c65495343d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4729780
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>