fpstudy: Refactor upstart to detect fp-board and use parameters file
This should not change the parameters the study server was invoked
with prior to this change, for bloonchipper.

See the following original CLs that introduced the parameters:
* crrev.com/c/2980015
* crrev.com/c/2639215

We detect the fingerprint board at runtime using cros config
and read the correct parameters file for that board.

BUG=b:292152710, b:144061947
TEST=None

Cq-Depend: chromium:4730508
Change-Id: Ie7e8c2c97548e230d9e359fab82e6844437d4050
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4729779
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>