arc-setup: Allow arc-handle-upgrade to access virtio-blk /data
arc-handle-upgrade needs to read ARCVM's /data/system/packages.xml to
send upgrade metrics, remove /data on certain types of upgrades, etc.
However, the file (actually, the entire /data) is currently inaccessible
from the host side on virtio-blk /data ARCVM devices.

To address the issue, this CL lets arc-handle-upgrade mount virtio-blk
/data on /home/root/<hash>/android-data/data when there is either an LVM
application container or a Concierge disk image, so that packages.xml
can be read from the same path as virtio-fs /data.

Note that /data is not mounted on ARCVM until arcvm-post-login-services
is started (and thus arc-handle-upgrade is finished). Hence the host
side mount does not disturb the guest side mount and vice versa.

One caveat is that the mount fails (and corresponding error messages are
emitted) when either /home/root/<hash>/android-data/data does not exist
or the android-data device is not formatted yet. This happens on the
first boot after opt-in.

BUG=b:224666661
TEST=arc-handle-upgrade works on virtio-fs /data devices, Concierge disk
TEST=image /data devices, and LVM application container /data devices.
TEST=ARCVM /data is properly mounted and still works on the guest side
TEST=on both the initial boot and succeeding regular boots.

Change-Id: I327d32aac418312e37baffaf608c0e513b85e70c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4733411
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Yury Khmel <khmel@google.com>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>