cryptohome: Move MountHelper (Mounter) to namespace mounter
This is the capstone for the separation of Mount and MountHelper. This
CL renames MountHelper->Mounter which is what it actually does and
removes the inherited relation to MountInterface. The original
relationship was created to ease the transition from in-process to
out-of-process mounter which is now complete.

BUG=b:290897808
TEST=FEATURES=test emerge-zork cryptohome

Change-Id: I90a59853531c226f20bbb80ca6bf01c5b2323942
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4733211
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>