featured: Update OWNERS file.
Joel hasn't worked on the project in some time, and expressed that he's
focusing on other things now, so he doesn't have bandwidth to keep up
with changes to featured. So, remove him from OWNERS and add in iby so
that we have two people listed directly.

BUG=None
TEST=CQ

Change-Id: I83e26802351e1414f0507288c8b76eda84e78aa4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4734154
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Joel Fernandes <joelaf@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>