cryptohome: Drop support for non-migrating auth factor loading
Migration is rolled out, we're not rolling it back, so in general we
should probably drop support for the non-migrating and non-USS code
paths.

However, trying to rip it out of auth session entirely it a bit more
invasive than I really want to attempt, so this is more of a half
measure of ripping it out of AuthFactorManager and the loading of auth
factors. There's some other changes to be made here, to better make the
Add and Remove operations a bit more transactional, and that's easier to
do if the code here is simplified.

BUG=b:293923071
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ia3c23e90291df1b251a8c59f55ffb7a16a2dc442
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4734333
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>