cryptohome: Drop the USS migration feature flag
This drops the explicit feature flag for "enable USS migration". It does
not drop the "enable/disable USS" file-based flag, and in fact in a few
cases it even leverages in tests to replace disabling the migration.

It would be good to remove the ability to disable USS as well but
there's a lot more tests that need to be rewritten to fix that,
unfortunately. So this change just does the lighter change to get rid of
the feature flag. This does remove the ability to enable USS without
enabling migration but that is no longer a useful feature.

BUG=b:293923071
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I2d496397f9132b6bb2ebf5342f5846e29fa2775c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4738481
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>