debugd: Polish error message for packet captures
packet_capture command on Crosh is showing the helper executable path
when the user runs the command with wrong arguments (see attached bug
for the screenhot). It's confusing from the user perspective because
they shouldn't be seeing the internal executables the command is
running. Update the error message to omit the executable path and make
it more generic.

BUG=b:255958875
TEST=Run `packet_capture` command on Crosh and check the new usage
suggestion.

Change-Id: I4db0e8ff5a6946bbada63c3d05e8846370b67104
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4738556
Tested-by: Irem Uguz <iremuguz@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Irem Uguz <iremuguz@google.com>