secanomalyd: adds forbidden intersection UMA
Adds a metric for reporting the number of system processes that violate
the forbidden intersection principle:
`ChromeOS.AnomalousProcCout.ForbiddenIntersection`. The forbidden
intersection process list is updated every 30 seconds, but the process
count is sent only once per execution of secanomalyd, and only in the
logged-in state.

BUG=b:275758280
TEST=The metric is emitted correctly and only when the user is logged in.
TEST=All unit tests pass.

Change-Id: Ifcf24a73f0cf523c7c303f05778ca25054909c96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4722294
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>