metrics: Skip check when recording EnterpriseRollback metrics on disk
Allow metrics for the EnterpriseRollback project to be recorded on
disk without calling AreMetricsEnabled() check.

Rollback metrics are recorded from different daemons and states of
the device, including just after powerwash. In some occasions it is
not possible to check the value of the consent status so we need to
skip the call to this method to ensure metrics can always be recorded
in disk.

Metrics will be discarded by the consent check in chromium when they
are read to be sent to the server. Therefore, it is safe to skip the
check in this instance.

BUG=b:261850979
TEST=FEATURES=test emerge-amd64-generic metrics

Change-Id: I70beca415d3efa0dff49c06d4b00b25d547d2526
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4741573
Tested-by: Cris Guerrero Romero <crisguerrero@chromium.org>
Reviewed-by: Jong Ahn <jongahn@chromium.org>
Commit-Queue: Cris Guerrero Romero <crisguerrero@chromium.org>