cryptohome: Move PrepareForRemoval after session check
We can't remove an active user, and there's a check for it. But
currently PrepareForRemoval is called before the check, such that we'll
have remove partial state of the user already before reaching the check
if Remove user is called when the user is active.

This should not have caused issues in production as Chrome shouldn't be
removing users while they're logged-in.

BUG=b:294173742
TEST=cryptohome unittests

Change-Id: Ifcbc218382052bd3e318c14f68fc7c7012364b4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4741438
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Auto-Submit: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>