biod: Simplify ResetEntropy
The original ResetEntropy logic is quite complicated: it tries to
instantiate the same amount of biometrics managers and call the reset
entropy methods on them. However, practically only 1 manager will be
present in CrOS now. Even if we do want to be generic and support
multiple managers, the tool should probably connect to dbus and send
requests to do the reset. In addition, the unit of resetting entropy
should probably be the FP device itself, instead of managers. For
example, currently no matter biod instantiates a biometrics_manager or a
auth_stack_manager, the underlying FP device is the same.

Simplify the logic to wash entropy of the known FP device. This will
work for both the old flow and the new flow.

BUG=b:274560824
TEST=biod builds
TEST=pending manual verification

Change-Id: Iaebc08c15eb1c93c864d42988fc07b6f853c40e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4741440
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>