oobe_config: Ensure rollback metrics file cleanup
Extend rollback_cleanup to ensure the rollback metrics file is deleted
if is stale and the device is unowned. Move OOBE completion flag check
from oobe_conf_restore.conf to rollback_cleanup to ensure the right
cleanup path is followed based on the status of the device.

When rollback has finished and the device is re-owned, the rollback
metrics file is deleted. However, it is possible that after powerwash
the device does not get re-owned. If there is a reboot when the
device is unowned and the rollback metrics file has not been modified
for 15 days, we need to delete the file to avoid leaking information.

See detailed information in go/cros-rollback-metrics-file-permissions.

BUG=b:261850979
TEST=1) FEATURES=test emerge-amd64-generic oobe_config
     2) Manual tests in hatch:
     2a) Owned device and call rollback_cleanup
	- Without rollback data -> file is not deleted
        - With rollback data -> file and rollback data are deleted
	- With rollback data but no metrics file -> rollback data
	  is deleted
     2b) Not owned device and call rollback_cleanup
        - Rollback data is not deleted
	- Metrics file is not deleted if modified <= 15 days
	- Metrics file is deleted if modified > 15 days
	- If metrics file does not exist, nothing happens
     2c) Reboot
        - Owned device goes through 2a cleanup path
	- Not owned device goes through 2b cleanup path

Change-Id: Ic540e3ca2c16ac0adab3b3e5f66718930af32ed9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4743239
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Cris Guerrero Romero <crisguerrero@chromium.org>
Commit-Queue: Cris Guerrero Romero <crisguerrero@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>