hiberman: Don't rely on symlink to determine DM device path
DeviceMapper::device_path() tries to determine the path of a DM device
by dereferencing the symlink /dev/mapper/<dev_name>. This is not
reliable for devices that have just been created, the symlink is
added by udev, so there could be a delay.

Instead of following the symlink use 'dmsetup info' to determine
the minor device node of the DM device. With that we can know
the path.

BUG=b:294218252
TEST=hibernate works on brya

Change-Id: If27f6dde9b8c5823ece148755723b347f6eb10b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4743580
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>