usb_bouncer: Use command argument for busnum in USB session metric
Currently, the USB session metric in the USB bouncer uses a mix of
environment variables and sysfs files to construct a structured metric
event. In rare cases where multiple devices with the same VID/PID are
connected at the same time and assigned the same device number on
different buses, we will not be able to tell which "remove" metric to
associate with each "attach" metric. This CL resolves the issue by
updating the USB bouncer to use a command argument to read the busnum so
that it can be recorded in both attach and remove metrics.

BUG=b:294245227
TEST=FEATURES=test emerge-${BOARD} chromeos-base/usb_bouncer, manually
checked attach and remove metrics in /var/lib/metrics/structured/events
included busnum.

Change-Id: If7cc5761757415019ad9c52e21500c421226c081
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4749619
Tested-by: Jameson Thies <jthies@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>