shill: Destroy Recorder explicitly in MetricsTest
Verification on a mock object will happen on its destruction if it's not
triggered manually. Since recorder is a singleton, in the current code
it will only be reset in the SetUp() of the next time, which means if a
failure happens in test A, we will see the error message in the next
test of A, which is misleading.

BUG=b:293995835
TEST=enabled parallel unit tests following the instruction in $BUG, and
  then `FEATUERS="test" emerge-$BOARD shill`

Change-Id: I35f977d5e87bc57ea6304dfee146455333d3ebfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4747180
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Helen Huang <helenhuang@chromium.org>