usb_bouncer: Log USB error uevents from the kernel
Currently the USB bouncer will record a metric when it receives a USB
error event from the kernel, but it does not log the event. That makes
it difficult to see that an error occurred when reading system logs.
This CL updates the USB bouncer to log a message when it records a
metric on an error event received from the kernel.

BUG=b:294251681
TEST=FEATURES=test emerge-${BOARD} chromeos-base/usb_bouncer, manually
checked that the log was in /var/log/messages

Change-Id: Ia1e51302c9cc9416629c43e0520094681bc9af95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4749285
Tested-by: Jameson Thies <jthies@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>