libchromeos-rs: Move panic_handler into this crate
The original source is in:
https://github.com/google/crosvm/blob/main/base/src/sys/unix/panic_handler.rs

The only modification is the use of nix's memfd_create
instead of the crosvm SharedMemory wrapper.

The only behavior difference should be the lack of
MFD_NOEXEC_SEAL on supported systems. Which should be irrelevant
for the panic signatures.

BUG=b:293488266
TEST=CQ

Change-Id: I2f962efda16a47b227b18e611205c1410a4e56e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4749030
Tested-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Dennis Kempin <denniskempin@google.com>