vm_tools/chunnel: Use vsock crate instead of crosvm_base::vsock
The API is almost the same, just needs some minor adjustments.
This was the last usage of vhost, so we can remove the re-export
from libchromeos-rs as well.

The new implementation will call the FIONBIO ioctl, which has
to be added to the seccomp policies.
The arm32 recv syscall also has to be enabled for arm32 boards.

BUG=b:293488266
TEST=cros deploy chunnel to kevin. restart chunneld.
Start `python3 -m http.server` in crostini and access
localhost:8000 in Chrome.

Change-Id: I084b500fbc9ef7ea6344c9063bb3d6da235c2394
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4749032
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Dennis Kempin <denniskempin@google.com>