common-mk: platform2_test: Support nested user namespaces
This patch removes the check that rejects --strategy=unprivileged when
the UID is 0. This check prevents Alchemy from using the strategy
entirely because it runs everything in an unprivileged user namespace
with UID=GID=0. Instead, we add another check to avoid accidentally
exposing /sys as writable, for which the UID check was initially added.

See the following thread for the context:
https://crrev.com/c/4666549/14..28/common-mk/platform2_test.py#b577

BUG=b:262458823
TEST=platform2_test.py \
     --sysroot=/build/arm64-generic --strategy=unprivileged --user=root \
     --bind-mount-dev bash  # succeeds
TEST=sudo platform2_test.py \
     --sysroot=/build/arm64-generic --strategy=unprivileged --user=root \
     --bind-mount-dev bash  # fails
TEST=unshare --user --map-root-user platform2_test.py \
     --sysroot=/build/arm64-generic --strategy=unprivileged --user=root \
     --bind-mount-dev bash  # succeeds

Change-Id: Ifaccbe51ae11e694ab4887a7e525c5c43d15699a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4751539
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>