patchpanel: Allow static IPv6 configuration for ConnectNamespace
Add an option to use static IPv6 for ConnectNamespace. This patch does
not result in any behavioral change as static IPv6 is expected to always
be false by default.

This patch is only a D-Bus change and the newly added field is not use
anywhere with only this patch.

In a follow-up patch, ConnectNamespace is expected to also return its
created IPv6 addresses if static IPv6 is set to true.

BUG=b:290887135
TEST=FEATURES=test emerge-<BOARD> chromeos-base/system_api patchpanel \
     patchpanel-client dns-proxy system-proxy

Change-Id: Id4f2c3c2c3aeea00a3be981d78779337437ea4e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4764570
Tested-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>