power: Rework battery saver backlight dimming
The previous battery saver dimming conflicted with user-requested
explicit brightness preferences, which resulted in a "jump" when
the user adjusts the brightness after battery saver is enabled.
This patch reworks the implementation to move the layering so that
the brightness uses the explicit brightness instead of the internal
brightness to mitigate this.

*However*, there is a caveat with this implementation, which will be
addressed after clarification with the relevant policy owners.

BUG=b:293232802
TEST=Maximize brightness, enable battery saver, increase brightness

Change-Id: Ib1b0308a4705fb99bd223c60274a7922ce316d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4754719
Commit-Queue: Sangwhan Moon <sxm@google.com>
Auto-Submit: Sangwhan Moon <sxm@google.com>
Tested-by: Sangwhan Moon <sxm@google.com>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>