libbrillo: Do not expose protofile headers
device_policy_impl header was including headers from protofile package
and the size of a DevicePolicyImpl object depended on protofile package
headers size.

This caused undefined behavior when the protofile headers change,
causing a rebuild of libpolicy but not of any dependents of libpolicy.
Now, if the dependents rely on the size of the object the header
advertized and their header is outdated and the actual object is bigger,
initialization of the object will overwrite memory that it should not
touch.

This change does two things:
1) Decouple size of DevicePolicyImpl from size of protofiles by using
   pointers.
2) Do not expose any headers that may be out of date by removing the
   protofile header includes from device_policy_impl.h

BUG=b:294577904
TEST=FEATURES=test emerge-brya libbrillo

Change-Id: I6cd52f8f1d4cb13d77ae98606eaebbf39a03bd2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4754655
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Tested-by: Miriam Polzer <mpolzer@google.com>