shill: tethering: increase default tethering enable timeout to 45s
The tethering enabling operation may involve setting up a new
connection with the DUN specific APN, either in parallel to the
default connection (multiplexed) or instead of it (so default
disconnected and the DUN-specific one connected as default). In those
cases, the original 10s timeout may be too optimistic. The timeout is
updated to 45s, which gives the cellular connection logic much more
time to succeed.

Given that the user can now abort the tethering operation manually,
using such a long timeout should not affect the user interactions.

BUG=b:278026029
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I8c1687a7b7082ccbba2ce19204654eb11308c31d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4753417
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>