oobe_config: Support default case for event record
Fix code to support the default case when attempting to record
the rollback structured metrics.

Recording is called in the target image and it is possible that the
event tracked in the file is newer and not supported in the old
version.

Update code to allow reaching the default without crashing.

BUG=b:261850979
TEST=Manual test to ensure a newest rollback metrics file does not
     crash when RecordStrcuturedMetric is called.

Change-Id: I7cefbe33aaa6bc75930c70be898da9a9e618edf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4755531
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Cris Guerrero Romero <crisguerrero@chromium.org>
Tested-by: Cris Guerrero Romero <crisguerrero@chromium.org>