hiberman: Make volume manager a singleton
There should be only one instance of volume manager, so make it a
singleton. Since volume manager is stateless a RwLock can be used
to 'guard access', so the volume manager instance can be held
multiple times, which avoids having to release the instance or
to pass it to another function in a few places.

BUG=b:293476116
TEST=hiberman works on brya

Change-Id: I56e90c7c58dbd55adba3b0dbd8eba76fd573f016
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4759665
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>