cryptohome: Don't use PLOG with selinux calls
Logging with PLOG is intended for logging errors from calls that set
errno but unfortunately this does not seem to consistently work with the
calls provided by libselinux.

We see a decent amount of logs that report failures with an error of
"Success" because errno=0, which is confusing at best. It could actually
be even worse if there are cases where a non-zero errno is logged but
it's actually wrong.

I'm not sure exactly why libselinux produces these bad values but my
best guess is that its own logging (selinux_log) is using vfprintf which
does not preserve errno.

BUG=b:294383856
TEST=built & unit tested only

Change-Id: Idd4625a7dd649b8abcfbfa7b87ad503250d87780
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4763385
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>