shill: NetworkApplier uses NetworkConfig as input
This patch creates a single interface between Network and NetworkApplier
that takes NetworkConfig, NetworkPriority, and NetworkApplier::Area as
defined in this patch as parameters. Internally in NetworkApplier it
will analysis the Area and split the task into subtasks as needed. This
interface is planned to be migrated as-is into a RPC between shill and
patchpanel when patchpanel is going to take all the network
configuration responsibility.

No behavior change in this patch.

BUG=b:264963034
BUG=b:269401899
TEST=unit
TEST=tast run $(DUT) network.*Routing*

Change-Id: Idc7d8b715af1296b4accafe3c02fa80fc74eb59c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4763633
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>