secanomalyd: minijail in forbidden intersection
Adds a check to the forbidden intersection violation logic, to exempt
the minijail binary. If the process command is "minijail0" we do not
perform the forbidden intersection check, as it is normal for minijail
to be in the forbidden intersection.
Updates the test to verify the logic of forbidden intersection
detection.

BUG=b:295182299
TEST=Manually verified list of forbidden intersection processes.
TEST=All unit tests pass.

Change-Id: I1ae50aed522124ed33ed7261c90146236ebc5197
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4763270
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>