cryptohome: Stop generating .sum files
In theory these files allows us to detect corruption or bitrot in
durable files. In practice...it just generates a bunch of extra files
that aren't useful.

All that happens if a checksum fails to pass is that an ERROR goes into
the logfiles and then the .sum file gets rewritten with the new
checksum. So even on the off chance we get corruption, we'll never see
it unless we happen to get a logfile from the first time the corruption
is detected.

While I can see arguments for improving our robustness in the face of
failures here, without any hard data (and the current checksums do not
really provide that, they're just screaming into the void) it's hard to
justify the extra mess. Personally, I'd argue that if we're really
worried about bitrot then we should actually do something even better
than just checksums, like generating some error correcting codes (or
using a file encoding that incorporates them directly, rather than
generating additional files).

BUG=b:289896513
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I2ccd9821a7b3383a4b5bcf3a4dc529f01241d7bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4763556
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>