hiberman: Create DM snapshots for cryptohome volumes
So far no DM snapshots were created for cryptohome volumes in
preparation for resuming. The snapshots weren't necessary
because the users cryptohome was not mounted during the
authentication for resume. This is changing, as Chrome needs
to peek into the feature flags (stored on cryptohome) to
support hibernate on enterprise enrolled devices. Remove
crypto home from the denylist of DM snapshots.

BUG=b:293895107
TEST=no file system corruptions when resuming from hibernate on
  brya with CL:4767276 and CL:3700622

Change-Id: I19117cedcc9ffa803e3509597b97380d852a3019
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4771741
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>