cryptohome: Make the transactional Add more robust
Currently the AddAuthFactor operation does not really handle the case
where the USS in-memory structure is modified but the overall operations
fails (for whatever reason).

This change fixes this by adding the ability to take and restore a
snapshot of the USS, and modifying the AddAuthFactor process to restore
the USS if any individual operation fails.

BUG=b:289883542
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I93cf85f714267840379cb61ad0e1fe4c4537d1fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4775979
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>