cryptohome: Handle AddAuthFactor collisions differently
This changes makes it an explicit error (checked early on in the
AddAuthFactor function) to try and add an AuthFactor when one already
exists. In exchange, it changes the code on the USS side to clobber an
existing entry when persiting the key.

The main case where this could be relevant is if you have a partially
completed remove, where a factor was removed but somehow couldn't be
deleted from the USS. With the current behavior future attempts to add a
new factor with the same label would fail due to the USS collision.

The upfront check should still preserve the existing behavior that you
can't Add a factor on top of one that already exists (you need to either
remove it first, or use update). But it relies on the actual standard
that AuthSession uses for "exists" (i.e. it can be successfully loaded
into the AuthFactorMap) and not just the key existing in the USS.

BUG=b:289883542
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I434fb813a353888e28a9292ddc2d7ac92698e84d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4779004
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>