modemfwd: Enable modemfwd recovery for EM060 modem
New support for EM060 modem should include the ability to recover the
modem. This CL enables it, and also generalizes the check for supported
modems, as the list is growing longer.

This patch is validated by updating the helper manifest to map the new
modem's usb vid/pid and helper to the old modem's firmware-variant, and
then running cellular.modemfwdRecovery.stress as usual. This allows us
to test the full recovery flow even without a final variant and helper
manifest defined.

BUG=b:291960630
TEST=Used above steps to stress recovery mechanism

Change-Id: I6a0411c0a9db90abc55601048beb1bff19e12aab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4779008
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Madhav <madhavadas@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Daniel Winkler <danielwinkler@google.com>