vm_tools: Add perfetto to concierge (optionally)
Adds an optional dependency to concierge for the perfetto library.
Perfetto is used in chrome and chromeos for system-wide profiling, in
this case we can use it to track request timelines/contention.

This is part of improving concierge's responsiveness, see
go/concierge-concurrency.

When USE=tracing is applied, binary size (B) goes from 1213696 to
1628136, about 410KB. It is unchanged if tracing is off (compared
with before this CL).

BUG=b:294160898
TEST=USE=tracing emerge-* vm_host_tools ... deploy ...

Cq-Depend: chromium:4775906
Change-Id: Ib4428521dafd168d5e66b196cb332cc9de037432
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4778245
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>