swap_management: Start zram swap by D-Bus message
Zram swap was started by launching the swap_management binary with the
--swap_start option. The binary then exits immediately, which causes the
wrong device path to be displayed in /proc/swaps.

Since there is no upstart job dependency for swap_start, we can use a
D-Bus message to the actual swap_management daemon (which persists all
the time the DUT is up so the mount namespace will not be unmounted) to
start the zram swap. However, we still want to run swap_management to
stop the zram swap at shutdown to avoid a race condition with the
process_killer.

BUG=b:295910132
TEST=betty vm

Change-Id: I955c5df2bdf95fa13ffc93892e20a1394f3b8f99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4780739
Tested-by: Chun-Tse Shao <ctshao@google.com>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Chun-Tse Shao <ctshao@google.com>