cryptohome: Remove VaultKeyset-AddAuthFactor from tpm simulator test
This CL removes the VaultKeyset-AddAuthFactor from the
AuthSessionTpmSimulator test, adds the setup to add VaultKeyset
through the VaultKeyset APIs and reduce the test set by removing USS
rollback scenarios.

Going forward AddAuthFactor API should only be used for adding a USS
factor. Therefore we use VaultKeyset APIs to add the VaultKeysets
for the USS migration test cases.

In addition current test introduce VK and USS as an initial and final
storage option to cover all test scenarios with USS  migration and
rollback of the migration. We don't have the rollback option for
either USS ot USS migration anymore; thus we don't need to include
USS/VK as an initial and final storage options. This CL therefore also
removes the storage options and parametrised tests that cover both way
scenarios.

The cases that need to be tested now is where user has one or more
VaultKeysets and migrated to USS, and the possible scenarios for adding
CryptohomeRecovery factor during mid-migration and after migration.

This CL is within the effort of removing VaultKeyset-AddAuthFactor
support and the USS flags from the code.

BUG=b:295940908
TEST=FEATURES=test emerge-betty cryptohome

Change-Id: Id6a66ace83f0a0776e3a4632a312bbb0b153522b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4782029
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Betul Soysal <betuls@google.com>