vm_tools: concierge: Consolidate VM Sockets to const std::string&
The VM socket path doesn't change during the lifetime of a VM and was
being regenerated on demand every time GetVmSocketPath() was called.
This is inefficient.

Changes the internal storage of the VM socket path to be a const
std::string that is populated during VM initialization and all external
usage to be const std::string&.

BUG=b:286976104
TEST=cros_run_unit_tests, CQ

Change-Id: Ie5f415a01ff8d67e90c7a4e570234f42d2bd2145
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4778252
Commit-Queue: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Auto-Submit: Kameron Lutes <kalutes@chromium.org>
Tested-by: Kameron Lutes <kalutes@chromium.org>